re-organized Comments
